refactor for getStaticProps
